Fixed most const warnings and added pragmas to ignore false warnings

+++ src/tools/cardos-tool.c
@@ -671,7 +671,7 @@ static int cardos_format(const char *opt_startkey)
 		return 1;
 	}
 	if (apdu.resplen < 0x04) {
-		printf("expected 4-6 bytes form GET DATA for startkey data, but got only %u\n", apdu.resplen);
+		printf("expected 4-6 bytes form GET DATA for startkey data, but got only %u\n", (unsigned int ) apdu.resplen);
 		printf("aborting\n");
 		return 1;
 	}
@@ -1032,7 +1032,7 @@ static int cardos_change_startkey(const char *change_startkey_apdu)
 		return 1;
 	}
 	if (apdu.resplen < 0x04) {
-		printf("expected 4-6 bytes form GET DATA for startkey data, but got only %u\n", apdu.resplen);
+		printf("expected 4-6 bytes form GET DATA for startkey data, but got only %u\n", (unsigned int) apdu.resplen);
 		printf("aborting\n");
 		return 1;
 	}
@@ -1127,7 +1127,7 @@ change_startkey:
 		return 1;
 	}
 	if (apdu.resplen < 0x04) {
-		printf("expected 4-6 bytes form GET DATA for startkey data, but got only %u\n", apdu.resplen);
+		printf("expected 4-6 bytes form GET DATA for startkey data, but got only %u\n", (unsigned int) apdu.resplen);
 		printf("aborting\n");
 		return 1;
 	}
